requirements.txt
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.1.2
colorama==0.4.4
dnspython==2.2.1
email-validator==1.1.3
Flask==2.1.1
Flask-Login==0.6.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.22.3
pandas==1.4.2
passlib==1.7.4
pycparser==2.21
pyodbc==4.0.32
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.35
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0
